Fixing conflicting libc definitions

+++ include/c4/klib/string.h
@@ -1,9 +1,10 @@
 #ifndef _C4_KLIB_STRING_H
 #define _C4_KLIB_STRING_H 1
+#include <stddef.h>
 
-void *memcpy( void *dest, const void *src, unsigned n );
-void *memset(void *s, int c, unsigned n);
-unsigned strlen(const char *s);
-char *strncpy( char *dest, const char *src, unsigned n );
+void *memcpy(void *dest, const void *src, size_t n);
+void *memset(void *s, int c, size_t n);
+size_t strlen(const char *s);
+char *strncpy(char *dest, const char *src, size_t n);
 
 #endif
